// web/src/app/admin/interfaces/solicitud-registro.ts
export interface SolicitudRegistro {
    solicitud_registro: number;
    corporacion?: string;
    empresa?: string;
    nit?: string;
    direccion?: string;
    pais?: string;
    departamento?: string;
    municipio?: string;
    codigo_postal?: string;
    nombre?: string;
    telefono?: string;
    correo_electronico?: string;
    fhcreacion?: string;
    procesada?: number;
    fhprocesada?: string;
}